Add files via upload
This is Breast-Cancer-Dataset-Analysis-using-ML, in this I have total 32 columns out of which Diagnosis is a Dependent variable. Using other columns which are all parameters we can decide whether that patient has A benign tumor (B) or A malignant tumor (M). I have developed 3 models out of which one model has around90% of accuracy.